requirements.txt
Flask
torch
langchain
langchain-community
langchain_groq
faiss-cpu
transformers
sentence-transformers
python-dotenv
